@bwmarrin 2019-04-11
I love this :)

But, I'm wondering if you would mind helping me and changing it slightly so to avoid breaking the API?

I'm thinking that maybe we can leave Epoch as an int64.  Create an epoch variable in the node struct that is the right type, then in the NewNode func take the int64 epoch value and apply it to the node.epoch value.   Then on all Generate() code you can reference the node.epoch value. This also lets us not add an additional global value curTime.

Then the Time func will take a bit of changing but it's a deprecated func that I intend to replace as a member func off of the node struct.


Does that make sense?