@nishaad78 2019-04-11
Yes, it makes sense. I also like this better as it hides this ugly hack: `curTime.Add(epoch).Sub(curTime))`