@bwmarrin 2019-04-11
Just to add this in here, as an interesting note.

I ran the benchmarks before and after this change and there's some interesting changes.

Before:
```
BenchmarkGenerate-12                     5000000               243 ns/op               0 B/op          0 allocs/op
BenchmarkGenerateMaxSequence-12         20000000               113 ns/op               0 B/op          0 allocs/op
```

After: 
```
BenchmarkGenerate-12                     5000000               319 ns/op               0 B/op          0 allocs/op
BenchmarkGenerateMaxSequence-12         20000000                76.3 ns/op             0 B/op          0 allocs/op
```

I'm amazed the BenchmarkGenerateMaxSequence drop so much while the BenchmarkGenerate went up.  I'll dig into this more later  - or you're welcome to explore it as well.