@nishaad78 2019-04-11
Yes, I think the comparison logic in the if statement with my change was less efficient. I've update it and here are the results, much better now :)

Before:
```
BenchmarkGenerate-4                      5000000               244 ns/op               0 B/op          0 allocs/op
BenchmarkGenerateMaxSequence-4          20000000               113 ns/op               0 B/op          0 allocs/op
```
After:
```
BenchmarkGenerate-4                      5000000               243 ns/op               0 B/op          0 allocs/op
BenchmarkGenerateMaxSequence-4          20000000                66.6 ns/op             0 B/op          0 allocs/op
```
It is faster in my implementation in `BenchmarkGenerateMaxSequence` because of the optimized `time.Since` func.
After (using time.Now()):
```
BenchmarkGenerate-4                      5000000               243 ns/op               0 B/op          0 allocs/op
BenchmarkGenerateMaxSequence-4          20000000               111 ns/op               0 B/op          0 allocs/op
```